fix mission cmd(MAV_CMD_DO_CHANGE_SPEED), manual deceleration, add log

+++ libraries/AP_Mission/AP_Mission.h
@@ -90,8 +90,8 @@ public:
     // change speed command structure
     struct PACKED Change_Speed_Command {
         uint8_t speed_type;     // 0=airspeed, 1=ground speed
-        float target_ms;        // target speed in m/s, -1 means no change
-        float throttle_pct;     // throttle as a percentage (i.e. 1 ~ 100), 0 means no change
+        int16_t target_ms;        // target speed in m/s, -1 means no change
+        int16_t throttle_pct;     // throttle as a percentage (i.e. 1 ~ 100), 0 means no change
         float expected_elapsed_time;
     };
